remove unused variables

+++ run_test.cpp
@@ -117,10 +117,10 @@ int main()
 	// printf("total read time: %.20lf\n", result);
 	int time = 0;
 	for (int i = 1; i <= SIZE; i++) {
-		for(int j = 1; j <= 16; j++) {
-			if(i >= j && i % j == 0) {
+		for (int j = 1; j <= 16; j++) {
+			if (i >= j && i % j == 0) {
 				for(int k = 1; k <= 32; k++) {
-				printf("time: %ld /", 100*time);
+				printf("time: %f /", (double)100*time);
 				result += ssd -> event_arrive(WRITE, j*k, 1, (double)100*time++);
 				}
 			}
